Create README.md
A small physics project to find the smallest possible resultant of five vectors from a finite list of possible vectors to choose from. After an inital vector and the list of vectors are entered, all possible combinations of five are computed and the resultants are printed in order of ascending values of magnitude.